migrate from tap to mocha
After make-fetch-happen update GitHub Actions started failing. Updating
tap to the latest version and migrating to mocha for addon tests fixes
GitHub Action failures.
