+++ package.json
@@ -38,13 +38,14 @@
   },
   "devDependencies": {
     "bindings": "^1.5.0",
+    "mocha": "^10.2.0",
     "nan": "^2.14.2",
     "require-inject": "^1.4.4",
     "standard": "^14.3.4",
-    "tap": "^12.7.0"
+    "tap": "^16.3.4"
   },
   "scripts": {
     "lint": "standard */*.js test/**/*.js",
-    "test": "npm run lint && tap --timeout=600 test/test-*"
+    "test": "npm run lint && tap --no-coverage --timeout=600 test/tap-* && mocha --timeout=600000 test/mocha-*"
   }
 }
