+++ test/tap-configure-python.js
