+++ test/tap-find-accessible-sync.js
