+++ test/tap-find-node-directory.js
