+++ test/tap-find-python.js
