+++ test/tap-find-visualstudio.js
