+++ test/tap-install.js
