+++ test/tap-options.js
